Add a utils file to store helper functions.
