+++ src/components/Header/Header.js
@@ -1,5 +1,6 @@
 import './Header.css';
 import logo from '../../images/logo.png';
+import { Link } from 'react-router-dom'
 
 const Header = () => {
     return (
