+++ src/components/PetCard/PetCard.js
@@ -1,28 +1,12 @@
 import './PetCard.css';
-import cat from '../../images/cat.png';
-import dog from '../../images/dog.png';
-import fish from '../../images/fish.png';
-import bone from '../../images/bone.png';
+
 import { Link } from 'react-router-dom';
 import { useState } from 'react';
+import { checkIcon, checkLinkImage } from '../../utils';
 
 const PetCard = ({ id, name, owner, type }) => {
 
-  const checkIcon = (petType) => {
-    if (petType === 'Dog') {
-      return <img className='pet-icon' src={dog} alt="Dog Icon" />
-    } else if (petType === 'Cat'){
-      return <img className='pet-icon' src={cat} alt="Cat Icon" />
-    }
-  }
 
-  const checkLinkImage = (petType) => {
-    if (petType === 'Dog') {
-      return <img className='detail-link-icon' src={bone} alt="Dog Detail Link" />
-    } else if (petType === 'Cat'){
-      return <img className='detail-link-icon' src={fish} alt="Cat Detail Link" />
-    }
-  }
 
   return (
     <div className="pet-card">
